requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyMuPDF==1.23.8
httpx==0.24.1
groq==0.11.0
python-dotenv==1.0.0
requests==2.31.0